Change BenchmarkExecutor to reduce memory stress.
BenchmarkExecutor was retaining all returns from 'run'
to disable potential optimizations.
This was problematic because as GWT is optimized and
performed better, it produces more output which reduces
in turn the performance very significantly, especially
if the returned object is big (which is in most of the
times).

Changed the executor so it no longer keeps all results,
instead only the last result is preserved and stored in a
global variable. This significantly changes the results of
the many benchmarks.

Change-Id: If2b4ba5150828963f251a93e30f7cca0af0c595f